[refactor]: LadderGenerator 클래스에서 맵을 만들고 만들어진 맵으로 Ladder를 생성해서 반환하도록 수정
