+++ src/main/java/kr/codesquad/FIlterValue.java
@@ -1,0 +1,8 @@
+package kr.codesquad;
+
+import java.util.Optional;
+
+@FunctionalInterface
+public interface FIlterValue {
+    <T> Optional<T> filter();
+}
